WD-1488 : Support to Make Namespace Public (#95)
* Namespace endpoints refactor

* Support for public namespace - 1

* Support for public namespace - 2

* WD-1488 Add getPublicNamespace endpoint

https://github.com/workduck-io/mex/pull/95

+++ src/main/kotlin/com/workduck/service/SnippetService.kt
@@ -146,7 +146,7 @@ class SnippetService {
         togglePublicAccess(snippetID, workspaceID, version, 0)
     }
 
-    private fun togglePublicAccess(snippetID: String, workspaceID: String, version: Int, accessValue: Long){
+    private fun togglePublicAccess(snippetID: String, workspaceID: String, version: Int, accessValue: Int){
         when(version == -1 ){
             true -> {
                 val latestVersion = snippetRepository.getLatestVersionNumberOfSnippet(snippetID, workspaceID)
